+++ src/ingest_validation_tools/upload.py
@@ -15,7 +15,7 @@ from urllib.parse import urljoin, urlsplit
 import requests
 
 from ingest_validation_tools.enums import OtherTypes, Sample
-from ingest_validation_tools.error_report import ErrorDict, ErrorDictException, InfoDict
+from ingest_validation_tools.error_report import ErrorDict, InfoDict
 from ingest_validation_tools.plugin_validator import (
     ValidatorError as PluginValidatorError,
 )
@@ -86,6 +86,8 @@ class Upload:
         ]
         self.errors = ErrorDict()
         self.info = InfoDict()
+        self.get_errors_called: bool = False
+        self.get_info_called: bool = False
 
         self.get_app_context(app_context)
 
@@ -102,7 +104,7 @@ class Upload:
             }
 
         except PreflightError as e:
-            self.errors.preflight.append(str(e))
+            self.errors.preflight.value = str(e)
 
     #####################
     #
@@ -126,19 +128,17 @@ class Upload:
         ).strip()
         self.info.git = git_version
 
-        try:
-            tsvs = {
-                Path(path).name: {
-                    "Schema": sv.table_schema,
-                    "Metadata schema version": sv.version,
-                    "Directory schema version": sv.dir_schema,
-                }
-                for path, sv in self.effective_tsv_paths.items()
+        tsvs = {
+            Path(path).name: {
+                "Schema": sv.table_schema,
+                "Metadata schema version": sv.version,
+                "Directory schema version": sv.dir_schema,
             }
-            self.info.tsvs = tsvs
-        except PreflightError as e:
-            self.errors.preflight.append(str(e))
+            for path, sv in self.effective_tsv_paths.items()
+        }
+        self.info.tsvs = tsvs
 
+        self.get_info_called = True
         return self.info
 
     def get_errors(self, **kwargs) -> ErrorDict:
@@ -155,7 +155,7 @@ class Upload:
             return self.errors
 
         if not self.effective_tsv_paths:
-            self.errors.preflight.append("There are no effective TSVs.")
+            self.errors.preflight.value = "There are no effective TSVs."
             return self.errors
 
         # Collect errors
@@ -169,18 +169,19 @@ class Upload:
         # Pass in run_plugins bool to modify behavior.
         if self.run_plugins is None:  # default behavior
             if self.errors:  # errors found, skip
-                self.errors.plugin_skip = (
+                self.errors.plugin_skip.value = (
                     "Skipping plugins validation: errors in upload metadata or dir structure."
                 )
             else:  # no errors, run plugins
                 logging.info("Running plugin validation...")
-                self.errors.plugin = self._get_plugin_errors(**kwargs)
+                self._get_plugin_errors(**kwargs)
         elif self.run_plugins:
             logging.info("Running plugin validation...")
-            self.errors.plugin = self._get_plugin_errors(**kwargs)
+            self._get_plugin_errors(**kwargs)
         else:
             logging.info("Skipping plugin validation.")
 
+        self.get_errors_called = True
         return self.errors
 
     def get_app_context(self, submitted_app_context: Dict):
@@ -368,7 +369,7 @@ class Upload:
         rows = read_rows(Path(tsv_path), self.encoding)
         fields = rows[0].keys()
         if missing_fields := [k for k in constrained_fields.keys() if k not in fields].sort():
-            raise ErrorDictException(f"Missing fields: {missing_fields}")
+            raise Exception(f"Missing fields: {missing_fields}")
         url_errors = self._find_and_check_url_fields(rows, constrained_fields, schema)
         return url_errors
 
@@ -433,7 +434,7 @@ class Upload:
         if shared_dir_errors:
             self.errors.reference.update({"Shared Directory References": shared_dir_errors})
 
-    def _get_plugin_errors(self, **kwargs) -> dict:
+    def _get_plugin_errors(self, **kwargs):
         plugin_path = self.plugin_directory
         if not plugin_path:
             return {}
@@ -461,8 +462,7 @@ class Upload:
                 # We are ok with just returning a single error, rather than all.
                 errors["Unexpected Plugin Error"] = [e]
         for k, v in errors.items():
-            errors[k] = sorted(v)
-        return dict(errors)  # get rid of defaultdict
+            self.errors.plugin[k] = sorted(v)
 
     #################################
     #
@@ -607,6 +607,8 @@ class Upload:
         url_fields = {}
         check = {k: v for k, v in row.items() if k in constrained_fields}
         for check_field, value in check.items():
+            if check_field in self.check_fields and not self.globus_token:
+                raise Exception("No token received to check URL fields against Entity API.")
             if check_field in ["parent_sample_id", "parent_dataset_id"]:
                 url_fields[check_field] = value.split(",")
             else: