+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Removing references to deprecated Search API endpoint 'assayname'
 - Update Phenocycler directory schema
 - Update GeoMx NGS directory schema
+- Update shared upload check
 - Release Xenium
 
 ## v0.0.26