+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Release Xenium
 - Create Xenium directory schema
 - Error reporting changes, addition of summarized counts
+- Update Segmentation Mask link
 
 ## v0.0.26
 - Update GeoMx NGS directory schema